Support for media that is decidedly not production ready and a bunch mroe to-dos
After spending 5 hours figuring out what to do with media for this, just to write a few lines of code, this modern email fable comes to mind:

A woman calls in a plumber when her washing machine breaks down. The plumber arrives, studies the machine, then produces a hammer and gives it a hefty whack.

The washing machine starts working again and the plumber presents a bill for £200.
'Two hundred pounds?' says the woman. 'All you did was hit it with the hammer.'
So the plumber gives her an itemised bill: 'Hitting washing machine with a hammer - £5. Knowing where to hit it - £195.'